Improve averaging with flagged image plane
In some pipeline image an entire frequency plane may be flagged, in this case the imaged portion of the sky is populated with zeros, while the non-imaged portion is populated with NaNs. In this case the inverse variance calculation for this plane returns a NaN. Using a NaN value as a weight in the weighted average of all the planes results in a NaN output image.

This commit alters the inverse variance calculation to output a zero value for an image composed of all zeros, all NaNs or a combination of both zeros and NaNs which improves the image calculated as the average over all planes.