Fix a bug in inverse_variance and use wcs more efficiently
The bug in inverse_variance was causing the full 50 iterations to be used every
time since the variable `cut` was being incorrectly handled on line 231. I've changed
the variable name to 'keep' to be more informative and fixed the negation. The loop
usually terminates after about 5 or less iterations and this therefore speeds up the
function by a factor of about 10.

The wcs call on line 71 was returning the unused Spectral and Stokes axes of the image
WCS and therefore generating extra dimensions of unused information when determining the
world coordinates of the pixels. Limiting the WCS to just the 2 celestial axes lessens
the memory usage by two thirds in the call to `pixel_to_skycoord` and simplifis the code
a bit.